backend: Split more 30 year regulation tests to their own files
